// components/toast/types/types.ts
export type ToastTextSize =
  | 'largeTitle'
  | 'title'
  | 'title2'
  | 'title3'
  | 'headline'
  | 'caption'
  | 'caption2'
  | 'footnote'
  | 'subheadline'
  | 'callout'
  | 'body';